Link to Profile section at first use
The first use is in tagged bytes, but the meaningful first use is
changed here not only to link to the fuller definition, but also to
clarify that the interpretation cannot be arbitrary.

Given there are multiple references to "base CoRIM schema" in the
profile definition, I've renamed the "Full CoRIM CDDL" appendix to more
clearly denote the base schema concept, but without "schema" due to
Issue#208.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>