Remove validity-map from core and remove meta-map
This also adds an "aside" document to describe how a CoRIM issuer may
model their CoRIM validity through modern PKI practices.

The validity of a document should be determined by the signing key's
expiration. Any other form of validity should be an extension to the
base.

Since the meta-map only had the signer-map in it, I've collapsed that
indirection.

Closes Issue#222