Clarify untagged value comparison requirements
As stated, it appears that keys of measurement-values-map in this document are
given more leeway for specifying comparisons than profiles. I interpret the
text as saying any untagged value for a key that is not 0 through 14 must be
compared with binary equality. There is no MUST in the current language, so
that does not seem to be the intention. A profile should be allowed to specify
a comparison method as irregular as raw-value and raw-value-mask without the
need for registering a CBOR tag number with the IANA.

If a profile's comparison method will be common to multiple profiles that are
intended for general verifier consumption, then we just recommend that the
comparison gets registered with a CBOR tag to avoid having to duplicate language
and potentially very slightly deviate from the original meaning in different
profiles.

Signed-off-by: Dionna Glaze <drdeeglaze@gmail.com>