Change element to object-id where appropriate.
The word "element" is used in RFC8610 to refer to tree nodes of a parsed
CBOR object. A "measured element" as it was referred to here is not
about CBOR values but logical or physical devices/components/modules
that this commit generalizes to the word "object".

The mkey type is meant to identify which object is getting measured, so
rename the codepoint from measured-element-type-choice to
measured-object-id-type-choice.