Disambiguate endorsed-triples semantics.
Addresses Issue#251. It's unclear if this implicit condition is also
expected of the conditional-endorsement-series-triples and
conditional-endorsement-triples. The phasing in the appraisal section
seems to suggest that it should be. That would seem to mean that if you
want "unconditional" endorsements, you need to have a
reference-triple-record that has an empty measurement-values-map, but
that is ill-formed.

Co-authored-by: Ned Smith <ned.smith@intel.com>