Add any-of matching operator to wrap measurement values
When you have requirements that fields 1-100 must be certain values, and
field 0 may be one of many values when in the presence of those other
field values, the representation of the matching condition repeats the
same 100 certain values for each alternative for field 0. It's far
simpler to allow for measurement values to have independently flexible
measurement conditions when in that case.

This is still a weak pattern expression language, but the "any-of" use
case is a particularly commonly needed one.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>