Use key => value directly in measurement-map
Rename "authorized-by" to "__authorities" since authorized-by is more of
a policy concept. The authorities are just a matching condition,
although there is more to discuss about delegation in Issue#244.

Remove integrity-registers as a measurement value since the register
name can be gived directly in the $measured-element-type-choice and
reuse the digests codepoint.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>