Account for "activating" tags from CoBOM in theory of operation.
The CoBOM tag can help narrow a search space and make an attestation
result self-explanatory, such as "unsupported CoBOM" or "missing tags
x,y,z from CoBOM".
Without a CoBOM, we wanted to still ensure that the appraisal process
was repeatable. Add a SHOULD for tracking the head state of the
knowledge base of all input CoRIMs to an appraisal session.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>