Quantize inputs with a session initialization phase.
The text already refers to an appraisal session and an appraisal
context, but they are not introduced as concepts the same way the ACS
is. Add Context and Session as specific constructs that serve distinct
purposes.

I think the "CoRIMs are collected" text is a bit too vague, so I put all
input fetching in phase 0. Validation and transformation is in phase 1.

I found it weird that Evidence Collection was not part of input
gathering, so I've moved it.

This change clarifies the meaning of "active" and "inactive", and what
"requires CoBOMs" means. It's just whether all available tags are
included in the search space or only referenced ones are. A CoBOM is not
a bundle of tags the way a CoRIM is. A CoBOM must reference tags that
are already in the input set. It just flips them to being usable in the
appraisal procedure.

I did not provide CDDL for the session or context since I think that
needs to be sequenced after PR#299.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>