Drop top level tagging requirement
Tagged type choices are not typical.
I would go so far as to drop the 500 tag as the entrypoint to CoRIM
altogether. NVIDIA is creating CoRIMs this way, but they are using a
different content-type in the protected header. I think we can drop it
in an follow-up.

This patch drops

* the need to tag the type choice
* the extensibility of concise-rim-type-choice, since extensibility is
  governed by a profile, and the profile is not known at this point
  in parsing.
* the need to tag the signed corim, since it is a COSE-sign1 with an
  unambigiuous content-type, and COSE-sign1 already has its own tag.

Addresses Issue #333, but 500 removal is TBD.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>