Drop top level tagging requirement (#337)
* Drop top level tagging requirement

Tagged type choices are not typical.
I would go so far as to drop the 500 tag as the entrypoint to CoRIM
altogether. NVIDIA is creating CoRIMs this way, but they are using a
different content-type in the protected header. I think we can drop it
in an follow-up.

This patch drops

* the need to tag the type choice
* the extensibility of concise-rim-type-choice, since extensibility is
  governed by a profile, and the profile is not known at this point
  in parsing.
* the need to tag the signed corim, since it is a COSE-sign1 with an
  unambigiuous content-type, and COSE-sign1 already has its own tag.

Addresses Issue #333, but 500 removal is TBD.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>

* Update corim.cddl

Fixed PR so that it builds.

* Change tags and content type to match errata1

Tagged type choices are not typical.
I would go so far as to drop the 500 tag as the entrypoint to CoRIM altogether. NVIDIA is creating CoRIMs this way, but they are using a different content-type in the protected header. I think we can drop it in an follow-up.

This patch drops

the need to tag the type choice
the extensibility of concise-rim-type-choice, since extensibility is governed by a profile, and the profile is not known at this point in parsing.
the need to tag the signed corim, since it is a COSE-sign1 with an unambigiuous content-type, and COSE-sign1 already has its own tag.
Addresses Issue #333, but 500 removal is TBD.

* Fix CoRIM examples with CDDL changes.

* Create signed example in Makefile.

* Remove signing from Makefile

CI doesn't have envsubst or openssl tools.

* Change tags and content type to match errata1

Tagged type choices are not typical.
I would go so far as to drop the 500 tag as the entrypoint to CoRIM altogether. NVIDIA is creating CoRIMs this way, but they are using a different content-type in the protected header. I think we can drop it in an follow-up.

This patch drops

the need to tag the type choice
the extensibility of concise-rim-type-choice, since extensibility is governed by a profile, and the profile is not known at this point in parsing.
the need to tag the signed corim, since it is a COSE-sign1 with an unambigiuous content-type, and COSE-sign1 already has its own tag.
Addresses Issue #333, but 500 removal is TBD.

* Fix CoRIM examples with CDDL changes.

* Create signed example in Makefile.

* Remove signing from Makefile

CI doesn't have envsubst or openssl tools.

* Add rim+cose with CoAP registration and magic numbers

* Update draft-ietf-rats-corim.md

fix broken section reference

* Update cddl/examples/payload-corim-4.diag

Co-authored-by: Thomas Fossati <tho.ietf@gmail.com>

* Remove spurious cbor-tags addition.

* Update rim+cose application copypasta

Co-authored-by: Thomas Fossati <tho.ietf@gmail.com>

* Update +cose magic number to #6.18, 4-elm array

Co-authored-by: Thomas Fossati <tho.ietf@gmail.com>

* Update draft-ietf-rats-corim.md

Co-authored-by: Andrew Draper <andrew.draper@altera.com>

* Update draft-ietf-rats-corim.md

Co-authored-by: Dionna Amalie Glaze <drdeeglaze@gmail.com>

---------

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>
Co-authored-by: nedmsmith <ned.smith@intel.com>
Co-authored-by: Thomas Fossati <tho.ietf@gmail.com>
Co-authored-by: Yogesh Deshpande <yogesh.deshpande@arm.com>
Co-authored-by: Andrew Draper <andrew.draper@altera.com>