Require at most one signer in corim-entity-map (#422)
The corim-signer-map identifies a single entity as the signer.
When only using unsigned CoRIMs and depending on the entities, there
isn't a structural limitation to stating multiple entities as signers.

Unless we're saying that every hop through nodes requires adding oneself
to the entities as manifest-signer, then we need to make that clear and
change the text that the manifest-signer is the one that originated the
CoRIM.